requirements.txt
smolagents
google-search-results
python-dotenv
requests